+++ src/config/defaults.py
@@ -7,6 +7,8 @@ DEFAULT_CONFIG = {
         'language': 'system',
         'theme': 'system',
         'deck_count': 10,
+        'remember_last_project': False,
+        'last_project_file': '',
     },
     'Audio': {
         'output_device': 'default',
@@ -155,9 +157,9 @@ VALID_DECK_RANGE = [1, 128]
 
 # Application info
 APP_NAME = 'MultiDeck Audio Player'
-APP_VERSION = '0.7.5'
+APP_VERSION = '0.8.0'
 APP_BUNDLE_ID = 'com.multideck.audioplayer'
-APP_CODE_NAME = 'Kampgarten'
+APP_CODE_NAME = 'Regentonne'
 APP_AUTHOR = 'Steffen Schultz'
 APP_WEBSITE = 'https://m45.dev'
 APP_LICENSE = 'MIT License'