+++ changelog.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 
 * Replace three dots with Ellipsis in menus
-* Implemented a project properties dialog to override automation settings during runtime
+* Implemented a project properties dialog to override automation settings during runtime (accessible via file menu).
+* An option to remember the last project on startup has been added to general settings.
 * Various code improvements
 
 ## [v0.7.5 (Kampgarten)] - 2026-07-24